2021-04-30 ScrambledBugs

+++ ScrambledBugs/ScrambledBugs/Fixes/SpeedMultUpdates.cs
@@ -2,7 +2,7 @@
 {
 	internal class SpeedMultUpdates
 	{
-		protected enum ActorValues : System.UInt32
+		protected enum ActorValues : System.Int32
 		{
 			SpeedMult = 30
 		}
@@ -46,7 +46,7 @@
 				IncludeLength = 5,
 				After = cpuRegisters =>
 				{
-					var speedMultUpdateListenerAddress = NetScriptFramework.Main.GameInfo.GetAddressOf(517376, (System.Int32)(0x8 * (System.UInt32)SpeedMultUpdates.ActorValues.SpeedMult), 0, "00 00 00 00 00 00 00 00"); // SkyrimSE.exe + 0x2F39A40
+					var speedMultUpdateListenerAddress = NetScriptFramework.Main.GameInfo.GetAddressOf(517376, 0x8 * (System.Int32)SpeedMultUpdates.ActorValues.SpeedMult, 0, "00 00 00 00 00 00 00 00"); // SkyrimSE.exe + 0x2F39A40
 					NetScriptFramework.Memory.WritePointer(speedMultUpdateListenerAddress, SpeedMultUpdateListenerAllocation.Address);
 				},
 			});
